改善remember me

+++ src/main/java/edu/fjnu/serviceImpl/PersistentLoginsServiceImpl.java
@@ -2,6 +2,7 @@ package edu.fjnu.serviceImpl;
 
 import edu.fjnu.entity.PersistentLogins;
 import edu.fjnu.mapper.PersistentLoginsMapper;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.authentication.rememberme.PersistentRememberMeToken;
@@ -10,6 +11,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
+
+import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 
 /**
  * @author xiaoze
@@ -28,37 +31,47 @@ public class PersistentLoginsServiceImpl implements PersistentTokenRepository {
     @Override
     public void createNewToken(PersistentRememberMeToken token) {
         PersistentLogins persistentLogins =new PersistentLogins();
-        persistentLogins.setUsername(token.getUsername());
-        persistentLogins.setSeries(token.getSeries());
-        persistentLogins.setToken(token.getTokenValue());
-        persistentLogins.setLastUsed(token.getDate());
-        persistentLoginsMapper.insertSelective(persistentLogins);
+        if(token!=null) {
+            persistentLogins.setUsername(token.getUsername());
+            persistentLogins.setSeries(token.getSeries());
+            persistentLogins.setToken(token.getTokenValue());
+            persistentLogins.setLastUsed(token.getDate());
+            persistentLoginsMapper.insertSelective(persistentLogins);
+        }
     }
 
     @Override
     public void updateToken(String series, String tokenValue, Date lastUsed) {
         PersistentLogins persistentLogins =new PersistentLogins();
-        persistentLogins.setSeries(series);
-        persistentLogins.setToken(tokenValue);
-        persistentLogins.setLastUsed(lastUsed);
-        persistentLoginsMapper.updateByPrimaryKeySelective(persistentLogins);
+        if(StringUtils.isNotEmpty(series))
+            persistentLogins.setSeries(series);
+
+        if(StringUtils.isNotEmpty(tokenValue))
+            persistentLogins.setToken(tokenValue);
+
+        if(lastUsed!=null)
+            persistentLogins.setLastUsed(lastUsed);
+        if(persistentLogins!=null)
+            persistentLoginsMapper.updateByPrimaryKeySelective(persistentLogins);
     }
 
     @Override
     public PersistentRememberMeToken getTokenForSeries(String seriesId) {
-        PersistentLogins persistentLogins =persistentLoginsMapper.selectByPrimaryKey(seriesId);
 
+        if(persistentLoginsMapper.selectByPrimaryKey(seriesId)!=null) {
+            PersistentLogins persistentLogins = persistentLoginsMapper.selectByPrimaryKey(seriesId);
 
-        if(persistentLogins!=null)
             return new PersistentRememberMeToken(persistentLogins.getUsername(),
                     persistentLogins.getSeries(),
                     persistentLogins.getToken(),
                     persistentLogins.getLastUsed());
+        }
         return null;
     }
 
     @Override
     public void removeUserTokens(String username) {
-        persistentLoginsMapper.deleteByUsername(username);
+        if(StringUtils.isNotEmpty(username))
+            persistentLoginsMapper.deleteByUsername(username);
     }
 }
